Update reviewers on DraftPublishedEvent
It is requested by our Gerrit users that reviewers are added only
for non-draft changes. This can be partially achieved by

[filter "-status:draft"]

However, when the draft is published the reviewers is not added
automatically.

To work this around, we suggested our users to create another
patchset and then the reviewers will be added.

This change addresses this issue. It will update reviewers on
DraftPublishedEvent too.

Change-Id: I9693603846c17c78179688c0614d8fa5b88f7df1